requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==5.0.6
django-appconf==1.0.6
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-imagekit==5.0.0
django-js-asset==2.2.0
djangorestframework==3.15.1
pilkit==3.0
pillow==10.3.0
psycopg2==2.9.9
python-dotenv==1.0.1
python-memcached==1.62
sqlparse==0.5.0
tzdata==2024.1
Unidecode==1.3.8
